requirements.txt
django
djongo5
pymongo
dnspython
djangorestframework
python-dotenv
djangorestframework-simplejwt
drf-yasg
pytest
pytest-django
django-cors-headers
black
isort
